removed datatype

+++ transform.py
@@ -28,7 +28,7 @@ def transform():
     print('====================================')
     
     # todo: change datatype 
-    df["Rating"]=df["Rating"].astype(int)
+   
     
     # drop unwanted columns
     df = df.drop(['hawa'], axis=1)
@@ -36,6 +36,7 @@ def transform():
     
     # todo: Null handle 
     df["desc"] = "No description"
+
     
     
     return df
